test: split out transaction admin tests and cover retry/transport edges
